+++ Weekand/.swiftlint.yml
@@ -13,8 +13,3 @@ disabled_rules:
 
 excluded:
 - Pods
-- Weekand/AppDelegate.swift
-- Weekand/SceneDelegate.swift
-
-included:
-- Weekand
