--- Weekand/Supporting Files/Preview.swift
@@ -1,51 +1,0 @@
-//
-//  Preview.swift
-//  Weekand
-//
-//  Created by 이호영 on 2022/05/20.
-//
-
-import SwiftUI
-
-enum DeviceType {
-    case iPhoneSE2
-    case iPhone8
-    case iPhone12Mini
-    case iPhone12Pro
-    case iPhone12ProMax
-
-    func name() -> String {
-        switch self {
-        case .iPhoneSE2:
-            return "iPhone SE"
-        case .iPhone8:
-            return "iPhone 8"
-        case .iPhone12Mini:
-            return "iPhone 12 Mini"
-        case .iPhone12Pro:
-            return "iPhone 12 Pro"
-        case .iPhone12ProMax:
-            return "iPhone 12 Pro Max"
-        }
-    }
-}
-
-#if canImport(SwiftUI) && DEBUG
-extension UIViewController {
-
-    private struct Preview: UIViewControllerRepresentable {
-        let viewController: UIViewController
-
-        func makeUIViewController(context: Context) -> UIViewController {
-            return viewController
-        }
-
-        func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
-        }
-    }
-
-    func showPreview(_ deviceType: DeviceType = .iPhone12Pro) -> some View {
-        Preview(viewController: self).previewDevice(PreviewDevice(rawValue: deviceType.name()))
-    }
-}
-#endif
