--- Weekand/Weekand/Presentation/SignScene/SignUp/ViewModel/SignUpViewModel.swift
@@ -1,9 +1,0 @@
-//
-//  SignUpViewModel.swift
-//  Weekand
-//
-//  Created by Daegeon Choi on 2022/05/20.
-//
-
-import Foundation
-
