--- Weekand/Weekand/Presentation/Utils/DesignSystem/ColorSet.swift
@@ -1,26 +1,0 @@
-//
-//  Color.swift
-//  Weekand
-//
-//  Created by Daegeon Choi on 2022/05/20.
-//
-
-import UIKit
-
-extension UIColor {
-    // MARK: Custom System Colors
-    static let mainColor = UIColor(hex: "#5086FF")
-    static let subColor = UIColor(hex: "#5086FF0A")
-    static let backgroundColor = UIColor.white
-    
-    static let accentColor = UIColor(hex: "#17191C")
-    static let lightGray = UIColor(hex: "#ECEFF2")
-    static let darkGray = UIColor(hex: "#AAAEB6")
-    
-    static let gray900 = UIColor(hex: "#17191C")
-    static let gray800 = UIColor(hex: "#32353A")
-    static let gray700 = UIColor(hex: "#515459")
-    static let gray600 = UIColor(hex: "#747880")
-    static let gray500 = UIColor(hex: "#8B929C")
-    static let gray400 = UIColor(hex: "#AAAEB6")
-}
