--- Weekand/Weekand/Presentation/Utils/DesignSystem/CommonShape.swift
@@ -1,16 +1,0 @@
-//
-//  CommonShape.swift
-//  Weekand
-//
-//  Created by Daegeon Choi on 2022/05/23.
-//
-
-import UIKit
-
-// View Spec
-let defaultFontSize: CGFloat = 13.5
-let defaultCornerRadius: CGFloat = 12
-
-// Constraint Spec
-let defaultSpacing: CGFloat = 30
-let defaultWidthDivider: CGFloat = 1.2
